# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/orders": {
      "get": {
        "operationId": "AdminOrdersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "paymentStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of all orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}/timeline": {
      "get": {
        "operationId": "AdminOrdersController_findTimeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order audit timeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}/status": {
      "patch": {
        "operationId": "AdminOrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}/payment-status": {
      "patch": {
        "operationId": "AdminOrdersController_updatePaymentStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}/tracking": {
      "patch": {
        "operationId": "AdminOrdersController_updateTracking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrackingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tracking information updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}/assign": {
      "patch": {
        "operationId": "AdminOrdersController_assignOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order assigned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}/notes": {
      "patch": {
        "operationId": "AdminOrdersController_updateNotes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderNotesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order notes updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}/cancel": {
      "patch": {
        "operationId": "AdminOrdersController_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/{id}/refund": {
      "patch": {
        "operationId": "AdminOrdersController_refundOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order refunded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/admin/orders/bulk/status": {
      "post": {
        "operationId": "AdminOrdersController_bulkUpdateStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
...
    "/checkout/guest/orders/{token}": {
      "get": {
        "operationId": "GuestCheckoutController_getOrder",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "checkout"
        ]
      }
    },
...
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
    "/custom-designs/admin/{id}/status": {
      "patch": {
        "operationId": "CustomDesignsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Custom design status updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update custom design status (admin)",
        "tags": [
          "custom-designs"
        ]
      }
    },
...
    "/admin/refunds/orders/{orderId}": {
      "get": {
        "operationId": "RefundsController_findByOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Refunds for an order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-refunds"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-commerce API",
    "description": "API para tienda y panel administrativo",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AuditLogEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "entity": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuditLogUserDto"
          },
          "before": {
            "type": "object"
          },
          "after": {
            "type": "object"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "action",
          "entity",
          "entityId",
          "createdAt"
        ]
      },
      "AuditLogListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogEntryDto"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "page",
              "limit",
              "total"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "AdminOrderUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "AdminOrderProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "AdminOrderProductVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/AdminOrderProductDto"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "AdminOrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "number"
          },
          "productionStatus": {
            "type": "string"
          },
          "productVariant": {
            "$ref": "#/components/schemas/AdminOrderProductVariantDto"
          },
          "customDesign": {
            "$ref": "#/components/schemas/AdminOrderCustomDesignDto"
          }
        },
        "required": [
          "id",
          "type",
          "quantity",
          "unitPrice",
          "productionStatus"
        ]
      },
      "AdminOrderListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "paymentStatus": {
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "shippingCost": {
            "type": "number"
          },
          "discountAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "trackingNumber": {
            "type": "string"
          },
          "carrier": {
            "type": "string"
          },
          "trackingUrl": {
            "type": "string"
          },
          "shippedAt": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AdminOrderUserDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderItemDto"
            }
          }
        },
        "required": [
          "id",
          "status",
          "paymentStatus",
          "subtotal",
          "shippingCost",
          "discountAmount",
          "totalAmount",
          "createdAt",
          "updatedAt",
          "user",
          "items"
        ]
      },
      "AdminOrderListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderListItemDto"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "page",
              "limit",
              "total"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "AdminOrderAddressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "addressLine1",
          "city",
          "state",
          "postalCode",
          "country",
          "phone"
        ]
      },
      "AdminOrderCouponDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "discountType": {
            "type": "string"
          },
          "discountValue": {
            "type": "number"
          }
        },
        "required": [
          "code",
          "discountType",
          "discountValue"
        ]
      },
      "AdminOrderPaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider",
          "amount",
          "status",
          "createdAt"
        ]
      },
      "AdminOrderDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "paymentStatus": {
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "shippingCost": {
            "type": "number"
          },
          "discountAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "trackingNumber": {
            "type": "string"
          },
          "carrier": {
            "type": "string"
          },
          "trackingUrl": {
            "type": "string"
          },
          "shippedAt": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AdminOrderUserDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderItemDto"
            }
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/AdminOrderAddressDto"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/AdminOrderAddressDto"
          },
          "coupon": {
            "$ref": "#/components/schemas/AdminOrderCouponDto"
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderPaymentDto"
            }
          },
          "assignedTo": {
            "$ref": "#/components/schemas/AdminOrderUserDto"
          },
          "adminNotes": {
            "type": "string"
          },
          "cancellationReason": {
            "type": "string"
          },
          "cancelledAt": {
            "type": "string"
          },
          "refundReason": {
            "type": "string"
          },
          "refundedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "paymentStatus",
          "subtotal",
          "shippingCost",
          "discountAmount",
          "totalAmount",
          "createdAt",
          "updatedAt",
          "user",
          "items",
          "payments"
        ]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PAID",
              "REFUNDED",
              "PENDING_PAYMENT",
              "IN_PRODUCTION",
              "READY_TO_SHIP",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdatePaymentStatusDto": {
        "type": "object",
        "properties": {
          "paymentStatus": {
            "type": "string",
            "enum": [
              "FAILED",
              "PENDING",
              "AUTHORIZED",
              "PAID",
              "REFUNDED"
            ]
          }
        },
        "required": [
          "paymentStatus"
        ]
      },
      "UpdateTrackingDto": {
        "type": "object",
        "properties": {
          "trackingNumber": {
            "type": "string"
          },
          "carrier": {
            "type": "string"
          },
          "trackingUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "AssignOrderDto": {
        "type": "object",
        "properties": {
          "assignedToId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UpdateOrderNotesDto": {
        "type": "object",
        "properties": {
          "adminNotes": {
            "type": "string"
          }
        }
      },
      "CancelOrderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
      "RefundOrderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "reason"
        ]
      },
      "BulkOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PAID",
              "REFUNDED",
              "PENDING_PAYMENT",
              "IN_PRODUCTION",
              "READY_TO_SHIP",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ]
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "status",
          "ids"
        ]
      },
...
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "shippingCost": {
            "type": "number"
          },
          "discountAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "paymentStatus": {
            "type": "string"
          },
          "trackingNumber": {
            "type": "string"
          },
          "carrier": {
            "type": "string"
          },
          "trackingUrl": {
            "type": "string"
          },
          "shippedAt": {
            "format": "date-time",
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "shipments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShipmentResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "estimatedDeliveryDate": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "subtotal",
          "shippingCost",
          "discountAmount",
          "totalAmount",
          "paymentStatus",
          "items",
          "shipments",
          "createdAt"
        ]
      },
...
      "UpdateStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_REVIEW",
              "APPROVED",
              "REJECTED"
            ]
          },
          "rejectionReason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "RefundListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RefundResponseDto"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "page",
              "limit",
              "total"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}